update register api
