+++ auth-service/auth-service/Program.cs
@@ -1,3 +1,5 @@
+using auth_service.Service;
+
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
@@ -8,6 +10,8 @@ builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
 
 builder.Services.AddHttpClient();
+
+builder.Services.AddScoped<AuthService>();
 
 var app = builder.Build();
