+++ auth-service/auth-service/appsettings.json
@@ -7,6 +7,8 @@
   },
   "AllowedHosts": "*",
   "KeyCloak": {
+    "adminEmail": "example@gmail.com",
+    "adminPass":  "1",
     "Address": "http://keycloak:8080/realms/Float-realm",
     "Register": "http://keycloak:8080/admin/realms/Float-realm/users",
     "Audience": "float-client",
